rag_server/requirements.txt
fastapi
uvicorn
openai==0.28
chromadb
pymongo
python-dotenv
